Consolidate ObsMapper versions (move tests)
